refactor(reporter): Migrate to new plugin API
Migrate the reporters to the new plugin API. With the reporter API
reporter implementations get their configuration options via the
`generateReport` function. With the new plugin API the configuration
options are provided as a constructor parameter. With this change both
ways are supported to allow for a step-by-step migration in the
following commits.

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.com>